NoxPanel/requirements.txt
# Core NoxPanel Backend Requirements
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
eventlet>=0.33.0
werkzeug>=2.3.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.11.0

# Authentication and Security
pyjwt>=2.8.0
bcrypt>=4.0.0
cryptography>=3.0.0

# HTTP and API
requests>=2.31.0
python-dotenv>=1.0.0
psutil>=5.9.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Enhanced Features
colorama>=0.4.6
click>=8.1.0
rich>=13.0.0
